locale: review Spanish translation to make it gender neutral
This patch reviews the Spanish translation to make it gender neutral.
With it, we make use of an inclusive language instead of the traditional
default male gender to refer to both males and females.

We use "persona usuaria" in places where we used "usuario", if possible.
Otherwise, we user "usuario/a". It may sound a bit repetitive but we
believe that using the traditional male gender implicitly excludes
references to females.

And, after all, it does not harm anyone to read the extra word "persona"
on the screen, the effort is worth for our purposes.

This patch also includes some minor nitpicks that I found while making
the translation.